# cmd/bnsapi/docs/swagger.yaml
definitions:
  account.Account:
    properties:
      certificates:
        items:
          items:
            type: integer
          type: array
        type: array
      domain:
        description: Domain references a domain that this account belongs to.
        type: string
      metadata:
        $ref: '#/definitions/weave.Metadata'
        type: object
      name:
        type: string
      owner:
        $ref: '#/definitions/weave.Address'
        description: |-
          Owner is a weave.Address that controls this account. Can be empty.

          An account can be administrated by the domain admin. In addition,
          ownership can be assigned to an address to allow another party to manage
          selected account.
        type: object
      targets:
        items:
          $ref: '#/definitions/account.BlockchainAddress'
        type: array
      valid_until:
        description: |-
          Valid until defines the expiration date for the account. Expired account
          cannot be used or modified. This date is always considered in context of
          the domain that this account belongs. Expired domain expires all belonging
          accounts as well, event if that account valid until date is not yet due.
        type: integer
    type: object
  account.BlockchainAddress:
    properties:
      address:
        description: |-
          An address on the specified blockchain network. Address is not a
          weave.Address as we cannot know what is the format of an address on the
          chain that this token instance links to. Because we do not know the rules
          to validate an address for any blockchain ID, this is an arbitrary bulk of
          data.
          It is more convenient to always use encoded representation of each address
          and store it as a string. Using bytes while compact is not as comfortable
          to use.
        type: string
      blockchain_id:
        description: An arbitrary blockchain ID.
        type: string
    type: object
...
  handlers.MultipleObjectsResponse:
    properties:
      objects:
        items:
          $ref: '#/definitions/util.KeyValue'
        type: array
    type: object
...
  orm.Model:
    type: object
...
  util.KeyValue:
    properties:
      key:
        $ref: '#/definitions/util.hexbytes'
        type: object
      value:
        $ref: '#/definitions/orm.Model'
        type: object
    type: object
  util.hexbytes:
    items:
      type: integer
    type: array
  weave.Address:
    items:
      type: integer
    type: array
  weave.Metadata:
    properties:
      schema:
        type: integer
    type: object
info:
  contact: {}
  license: {}
  title: BNSAPI documentation
paths:
  /account/accounts:
    get:
      description: |
        The list is either the list of all the starname (orkun*neuma) for a given premium starname (*neuma), or the list of all starnames for a given owner address.
        You need to provide exactly one argument, either the premium starname (*neuma) or the owner address.
      parameters:
      - description: 'Premium Starname ex: *neuma'
        in: query
        name: starname
        type: string
      - description: The owner address format is either in iov address (iov1c9eprq0gxdmwl9u25j568zj7ylqgc7ajyu8wxr)
          or hex (C1721181E83376EF978AA4A9A38A5E27C08C7BB2)
        in: query
        name: owner
        type: string
      - description: Query by domain
        in: query
        name: domain
        type: string
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MultipleObjectsResponse'
        "404": {}
        "500": {}
      summary: Returns a list of `bnsd/x/account` entities (like orkun*neuma).
      tags:
      - Starname
  /account/domains/:
    get:
      description: |-
        The list of all premium starnames for a given admin.
        If no admin address is provided, you get the list of all premium starnames.
      parameters:
      - description: The admin address may be in the bech32 (iov1c9eprq0gxdmwl9u25j568zj7ylqgc7ajyu8wxr)
          or hex (C1721181E83376EF978AA4A9A38A5E27C08C7BB2) format.
        in: query
        name: admin
        type: string
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MultipleObjectsResponse'
        "404": {}
      summary: Returns a list of `bnsd/x/domain` entities (like *neuma).
      tags:
      - Starname
...
  /account/resolve/{starname}:
    get:
      description: |-
        Resolve a given starname (like orkun*neuma) and return all metadata related to this starname,
        list of crypto-addresses (targets), expiration date and owner address of the starname.
      parameters:
      - description: 'starname ex: orkun*neuma'
        in: path
        name: starname
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/account.Account'
        "404": {}
        "500": {}
      summary: Resolve a starname (orkun*neuma) and returns a `bnsd/x/account` entity
        (the associated info).
      tags:
      - Starname
...
swagger: "2.0"